complete v1 endpoints

+++ taskly-fastapi/app/services/task.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 from app.model.base_model import Task as TaskModel
 from app.schema.schema import Task as TaskInterface
 
@@ -29,8 +30,27 @@ class TaskServices():
         self.db.refresh(new)
         return new
 
+    def put_task(self, idTask: int, task_data: TaskInterface):
+        task = self.get_task_by_id(idTask)
+
+        if not task:
+            return 'Not found'
+
+        task.title = task_data.title
+        task.description = task_data.description
+        task.status = task_data.status
+        task.updated = datetime.utcnow()
+
+        self.db.commit()
+        self.db.refresh(task)
+        return task
+
     def delete_task(self, idTask: int):
         task = self.get_task_by_id(idTask)
+
+        if not task:
+            return 'Not found'
+
         self.db.query(TaskModel).filter(TaskModel.id == idTask).delete()
         self.db.commit()
         return task
